Move login pages to seperate folder
